requirements.txt
django>=5
daphne>=4
channels>=4
djangorestframework>=3.15
djangochannelsrestframework>=1.2
channels-redis>=4
